Fix FormFile type
FormFile is returned to the web as part of FormPictureSelectResult, hence
it itself is a Result.
Since it will be converted to JSON as such, if not implementing Result it
will be missed by Proguard settings and be obfuscated (and the JSON keys
will be the obfuscated names instead of the original variable names...).